backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.13
python-dotenv==1.0.1
